devenv: Fully functional setup
Fixed a typo in the user.yml that was causing a 'group not found' error as well as fixing the fact gathering issue when looking up the mtu settings for the docker daemon.  Multiple apply/destroy cycles have been successful at this point.  This project can now be used by others.

Task: N/A
Testing: manual